Add separate file for `TagTci` command.
